refactoring old rudder code

+++ Modules/Calculation_Unit/calculation_unit.hpp
@@ -9,7 +9,7 @@ class Calculation_Unit
     public:
       double  degrees_to_radians(double degrees);
       double  radians_to_degrees(double radians);
-      double  convert_coordinates(double v_high, double v_low, double u_high, double u_low, double position);
+      int     convert_coordinates(double v_high, double v_low, double u_high, double u_low, double position);
       double  calculate_rudder_position(VEC2 vector);
       double  calculate_sail_position(VEC2 vector);
       VEC2    degrees_to_vector(double value);
